refactor: refactor pages sub components gf-233
