+++ apps/frontend/src/pages/auth/auth.tsx
@@ -15,7 +15,7 @@ import {
 	type UserSignUpRequestDto,
 } from "~/modules/users/users.js";
 
-import { SignInForm, SignUpForm } from "./components/components.js";
+import { SignInForm, SignUpForm } from "./libs/components/components.js";
 import styles from "./styles.module.css";
 
 const Auth = (): JSX.Element => {
