+++ apps/frontend/src/pages/auth/libs/components/sign-up-form/libs/constants/constants.ts
