+++ apps/frontend/src/pages/auth/libs/components/sign-up-form/sign-up-form.tsx
