+++ apps/frontend/src/pages/projects/libs/components/project-card/project-card.tsx
