+++ apps/frontend/src/pages/projects/libs/components/project-card/styles.module.css
