+++ apps/frontend/src/pages/projects/libs/components/project-create-form/libs/constants/constants.ts
