+++ apps/frontend/src/pages/projects/libs/components/project-create-form/libs/constants/default-project-create-payload.constant.ts
