+++ apps/frontend/src/pages/projects/libs/components/project-menu/project-menu.tsx
