+++ apps/frontend/src/pages/projects/libs/components/project-search/project-search.tsx
