+++ apps/frontend/src/pages/projects/libs/components/project-search/styles.module.css
